+++ keyoscacquire/keyoscacquire/__init__.py
@@ -4,7 +4,7 @@
 Andreas Svela // 2019
 """
 
-__version__ = '1.0.1'
+__version__ = '1.1.0'
 
 
 # Default options