+++ keyoscacquire/keyoscacquire/default_options.py
@@ -13,4 +13,3 @@ FILETYPE = ".csv"       # filetype of exported data, can also be txt/dat etc.
 EXPORT_PNG = True       # export png of plot of obtained trace
 SHOW_PLOT = False       # show each plot when generated (program pauses until it is closed)
 TIMEOUT = 15000         # ms timeout for the instrument connection
-DEBUG = False           # print more details