+++ keyoscacquire/keyoscacquire/installed_command_line_funcs.py
@@ -54,7 +54,7 @@ def single_connection_command_line():
 def single_trace_command_line():
     """Function installed on the command line: Obtains and stores a single trace."""
     parser = argparse.ArgumentParser(usage=acqprog.get_single_trace.__doc__)
-    parser.add_argument('-f', nargs='?', help='Specify filename base')
+    parser.add_argument('-f', nargs='?', help='Specify filename (without extension, \'.csv\' is added)')
     parser.add_argument('-a', nargs='?', help='Specify acquire type: {HRESolution, NORMal, AVER<m>} where <m> is the number of averages in range [1, 65536]')
     args = parser.parse_args()
 
@@ -65,7 +65,7 @@ def num_traces_command_line():
     parser = argparse.ArgumentParser(usage=acqprog.get_n_traces.__doc__)
     parser.add_argument('-f', nargs='?', help='Specify filename base')
     parser.add_argument('-a', nargs='?', help='Specify acquire type: {HRESolution, NORMal, AVER<m>} where <m> is the number of averages in range [1, 65536]')
-    parser.add_argument('-n', nargs='?', help='Specify number of traces to obtain')
+    parser.add_argument('-n', nargs='?', help='Specify number of traces to obtain (set to 1 if not specified)')
     args = parser.parse_args()
 
     acqprog.run_programme("num_traces", ['', args.f, args.a, args.n])