+++ keyoscacquire/docs/changelog.md
@@ -1,6 +1,8 @@
 # Changelog
 
-NEED TO INCORPORATE NEW FUNC IN README
-
-- v1.0.1: New command line function for obtaining a set number of traces consecutively, Added license, changes in README
+- v1.1.0: Extra command line function, logging enabled
+  - Added command line function for obtaining a given number of traces consecutively
+  - Former debugging print is now directed to logging.debug() 
+  - Added license file
+  - Changes in README
 - v1.0.0: First release